Adding groups
Save TrackLocalStaticRTP to forward rtps to. For each track a
PeerConnection receives it will create its respective
TrackLocalStaticRTP which can be added to other PeerConnection
connections to forward the rtp.
This commit only includes one group but it has enough capabilities added
to the WebRTCConnection class to handle different connections.