perfect negotiation instead of sync offer/answer
negotiation logic should be driven by the RTCPeerConnection itself,
meaning, that when a peer connection reports it needs negotiation then
the offer will be sent instead of creating it ourselves manually.

There is a small issue still present where if you start the
PeerConnection on the client fast enough, it will cause the addition of
ice candidates to fail because the RemoteDescription is yet to be added:

```
WARN signal_server/src/webrtc.rs:301: Error adding ice candidate ErrNoRemoteDescription
```

Clicking start again in the test client will succeed as the remote offer
is added by then.